test(agent): isolate RFC64 recovery lifecycles
