+++ packages/agent/vitest.rfc64-unit-tests.ts
@@ -59,6 +59,8 @@ export const RFC64_UNIT_TESTS = [
   "test/rfc64-private-catalog-activation-config-v1.test.ts",
   "test/rfc64-private-sender-key-roster.test.ts",
   "test/rfc64-rollout-authority.integration.test.ts",
+  "test/rfc64-replay-worklist.integration.test.ts",
+  "test/rfc64-catalog-replay-connection-runtime-v1.test.ts",
   "test/rfc64-private-catalog-gate-artifact.test.ts",
   "test/rfc64-policy-cell-v1.test.ts",
 ] as const;
