+++ packages/agent/vitest.unit.config.ts
@@ -192,8 +192,10 @@ export default defineConfig({
       "test/context-graph-join-policy.test.ts",
       "test/bounded-operation.test.ts",
       "test/private-read-chain-authority.test.ts",
+      "test/context-graph-authority-retry.test.ts",
       "test/context-graph-access-policy-state.test.ts",
       "test/context-graph-historical-name-binding.test.ts",
+      "test/context-graph-registration-resolution.test.ts",
       "test/context-graph-list-concurrency.test.ts",
       "test/cg-resolve-refresh.test.ts",
       "test/private-cg-membership-bootstrap.test.ts",
